requirements.txt
Django==4.1.3
python-dotenv==0.21.1
stripe==5.2.0
